Move DisplayGithubRepoData to DisplayGithubRepoData.swift
